Moved HeaderInfo to headers package
HeaderInfo will be required by the new thermal-writer binary (coming
soon).
