+++ headers/headerinfo_test.go
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-package main
+package headers
 
 import "github.com/TheCacophonyProject/go-cptv/cptvframe"
